refactor(server): extract spawn_structured_session from create_session handler
Move the session build/persist/upsert/ACP-spawn core of the create_session
HTTP handler into a reusable domain function so a future non-HTTP caller (a
cron scheduler) can create a structured-view ACP session without duplicating
the logic. The handler keeps all request decoding, validation, and response
construction: it decodes the request into a StructuredSessionSpec, calls
spawn_structured_session, and builds the HTTP response from the returned
SpawnOutcome exactly as before. Behavior-preserving; the panic-vs-build-error
status distinction is preserved via a SessionBuildPanicked marker error.
